// src/features/engine/modules/instrumentals-node/types/inst.category.type.ts
export type InstrumentType =
  | "piano"
  | "chromatic_percussion"
  | "organ"
  | "guitar_clean"
  | "guitar_nylon"
  | "guitar_jazz"
  | "guitar_overdriven"
  | "guitar_distortion"
  | "bass"
  | "string"
  | "ensemble"
  | "brass"
  | "reed"
  | "pipe"
  | "synth_lead"
  | "synth_pad"
  | "synth_effect"
  | "ethnic"
  | "percussive"
  | "sound_effects";